requirements.txt
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
black==23.9.1
click==8.1.7
crispy-bootstrap4==2022.1
decorator==5.1.1
Django==4.2.6
django-crispy-forms==2.0
django-dotenv==1.4.2
django-environ==0.11.2
executing==2.0.0
gunicorn==21.2.0
ipython==8.16.1
jedi==0.19.1
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
numpy==1.26.2
packaging==23.2
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
python-dotenv==1.0.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.11.2
wcwidth==0.2.8
whitenoise==6.5.0
